+++ src/Discord.Net.Core/Entities/Interactions/IDiscordInteraction.cs
@@ -112,6 +112,11 @@ namespace Discord
         GuildPermissions Permissions { get; }
 
         /// <summary>
+        ///     Gets the attachment size limit in bytes.
+        /// </summary>
+        ulong AttachmentSizeLimit { get; }
+
+        /// <summary>
         ///     Responds to an Interaction with type <see cref="InteractionResponseType.ChannelMessageWithSource"/>.
         /// </summary>
         /// <param name="text">The text of the message to be sent.</param>