+++ src/Discord.Net.Core/Entities/Interactions/MessageComponents/Builders/SelectMenuBuilder.cs
@@ -33,12 +33,12 @@ public class SelectMenuBuilder : IInteractableComponentBuilder
     public string CustomId
     {
         get => _customId;
-        set => _customId = value?.Length switch
+        set
         {
-            > ComponentBuilder.MaxCustomIdLength => throw new ArgumentOutOfRangeException(nameof(value), $"Custom Id length must be less or equal to {ComponentBuilder.MaxCustomIdLength}."),
-            0 => throw new ArgumentOutOfRangeException(nameof(value), "Custom Id length must be at least 1."),
-            _ => value
-        };
+            Preconditions.AtLeast(value.Length, 1, nameof(CustomId));
+            Preconditions.AtMost(value.Length, ComponentBuilder.MaxCustomIdLength, nameof(CustomId));
+            _customId = value;
+        }
     }
 
     /// <summary>
@@ -61,12 +61,12 @@ public class SelectMenuBuilder : IInteractableComponentBuilder
     public string Placeholder
     {
         get => _placeholder;
-        set => _placeholder = value?.Length switch
+        set
         {
-            > MaxPlaceholderLength => throw new ArgumentOutOfRangeException(nameof(value), $"Placeholder length must be less or equal to {MaxPlaceholderLength}."),
-            0 => throw new ArgumentOutOfRangeException(nameof(value), "Placeholder length must be at least 1."),
-            _ => value
-        };
+            Preconditions.AtLeast(value.Length, 1, nameof(Placeholder));
+            Preconditions.AtMost(value.Length, MaxPlaceholderLength, nameof(Placeholder));
+            _placeholder = value;
+        }
     }
 
     /// <summary>
@@ -109,7 +109,6 @@ public class SelectMenuBuilder : IInteractableComponentBuilder
         {
             if (value != null)
                 Preconditions.AtMost(value.Count, MaxOptionCount, nameof(Options));
-
             _options = value;
         }
     }