+++ src/Discord.Net.Rest/DiscordRestApiClient.cs
@@ -2844,6 +2844,11 @@ namespace Discord.API
                 query += $"&exclude_ended={args.ExcludeEnded.Value}";
             }
 
+            if (args.ExcludeDeleted.IsSpecified)
+            {
+                query += $"&exclude_deleted={args.ExcludeDeleted.Value}";
+            }
+
             return SendAsync<Entitlement[]>("GET", () => $"applications/{CurrentApplicationId}/entitlements{query}", new BucketIds(), options: options);
         }